backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
redis==5.2.0
web3==6.20.2
pydantic==2.9.2
python-dotenv==1.0.1
jsonschema==4.23.0
pytest==8.3.3
